+++ app/src/main/java/com/tovo/eat/ui/home/homemenu/kitchen/KitchenItemViewModel.java
@@ -123,7 +123,7 @@ public class KitchenItemViewModel {
     public void removeFavourite() {
 
         AppPreferencesHelper appPreferencesHelper = new AppPreferencesHelper(MvvmApp.getInstance(), AppConstants.PREF_NAME);
-        int userId = appPreferencesHelper.getCurrentUserId();
+        Long userId = appPreferencesHelper.getCurrentUserId();
         if (!MvvmApp.getInstance().onCheckNetWork()) return;
 
         if (favID != 0)
@@ -167,7 +167,7 @@ public class KitchenItemViewModel {
         //   AlertDialog.Builder builder=new AlertDialog.Builder(CartActivity.this.getApplicationContext() );
 
         AppPreferencesHelper appPreferencesHelper = new AppPreferencesHelper(MvvmApp.getInstance(), AppConstants.PREF_NAME);
-        int userId = appPreferencesHelper.getCurrentUserId();
+        Long userId = appPreferencesHelper.getCurrentUserId();
 
 
         try {
@@ -203,7 +203,7 @@ public class KitchenItemViewModel {
     }
 
     public interface KitchenItemViewModelListener {
-        void onItemClick(Integer id);
+        void onItemClick(Long id);
 
         void addFavourites(Integer id, String fav);
