+++ app/src/main/java/com/tovo/eat/ui/home/homemenu/kitchen/KitchenResponse.java
@@ -42,7 +42,7 @@ public class KitchenResponse {
 
         @SerializedName("makeituserid")
         @Expose
-        private Integer makeituserid;
+        private Long makeituserid;
         @SerializedName("makeitusername")
         @Expose
         private String makeitusername;
@@ -132,11 +132,11 @@ public class KitchenResponse {
             this.collection = collection;
         }
 
-        public Integer getMakeituserid() {
+        public Long getMakeituserid() {
             return makeituserid;
         }
 
-        public void setMakeituserid(Integer makeituserid) {
+        public void setMakeituserid(Long makeituserid) {
             this.makeituserid = makeituserid;
         }
